refactor(clipboard): replace golang.design/x/clipboard with go-nativeclipboard
go-nativeclipboard uses purego (no CGO) and has native FreeBSD support.
This eliminates the need for platform-specific clipboard files entirely.
